[Feat] #7 - Data model separation - Todo, Schedule, Today
